+++ libs/snackbars/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [2.0.2](https://github.com/cowprotocol/cowswap/compare/snackbars-v2.0.1...snackbars-v2.0.2) (2026-02-05)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/common-hooks bumped to 2.1.0
+    * @cowprotocol/ui bumped to 2.1.0
+
 ## [2.0.1](https://github.com/cowprotocol/cowswap/compare/snackbars-v2.0.0...snackbars-v2.0.1) (2026-02-02)